Restructure manage tokens, so chooseContract is the only part with token header
