+++ packages/browser-wallet/src/popup/pages/Account/Tokens/ManageTokens/utils.ts
@@ -22,7 +22,6 @@ export type DetailsLocationState = {
 };
 
 export const manageTokensRoutes = {
-    update: 'update',
     details: 'details',
 };
